+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -37,6 +37,7 @@
 #include "Rule18p7Check.h"
 #include "Rule18p8Check.h"
 #include "Rule19p2Check.h"
+#include "Rule3p1Check.h"
 #include "Rule8p10Check.h"
 #include "Rule8p11Check.h"
 #include "Rule8p14Check.h"
@@ -109,6 +110,8 @@ public:
         "misrac-rule-18p8");
     CheckFactories.registerCheck<Rule19p2Check>(
         "misrac-rule-19p2");
+    CheckFactories.registerCheck<Rule3p1Check>(
+        "misrac-rule-3p1");
     CheckFactories.registerCheck<Rule8p10Check>(
         "misrac-rule-8p10");
     CheckFactories.registerCheck<Rule8p11Check>(